Pulled Next Gen NFL Data
Next Gen data was pulled in a brute force manner by copying the data and pasting into a csv file. Then reading this into the explore_data.ipynb notebook. Scraping the site is difficult seeing as it requires a web crawler due to the dynamic rendering of the site.

Next step is to merge the combine, qb_drafts, and next_gen dataframes into one that can be used for analysis.